Extract the shared review surface into src/features/review
The anonymous /diff page reached into src/pages/Dashboard/ScanDetail/ for
filterDiffEntries, findingCountsByPath, RiskSignalsSection, and
FindingWithDiffStatus. Two sibling pages rendering the same review meant one
of them owned code the other depended on, so any change to the workbench's
internals silently belonged to the public surface too.

Moves what both surfaces use into src/features/review/ and leaves what is tied
to the persisted scan model (scanFilesToFileRecords, annotatePersistedFindings,
the DiffWorkbench state machine) in ScanDetail.

The shared finding type is now ReviewFinding — only the fields the review UI
renders — rather than a persisted scan row. The public diff previously had to
fabricate scanId: "" and ruleVersion to satisfy a component that never read
them; it no longer does. ReleaseRecommendation's local RecommendationFinding
was a near-duplicate of the same shape and collapses into it.

test/scan-detail-diff-helpers.test.mjs splits along the same seam, with the
shared helpers covered by test/review-diff-entries.test.mjs.

docs/ui.md and AGENTS.md record the rule the old layout violated: a page must
not import from another page's directory.
